Use Authorization header for authentication (#42)
* Use Authorization header for authentication

Using the access_token query parameter runs the risk of exposing the
access token in common stack traces. Switch to using an
"Authorization: Bearer token" header instead.

Some of the exceptions raised by requests include the url in the
exception text. If the stack trace is shared publicly somewhere, this
would lead to the exposure of an API token.

* run black

* Address review feedback

* Simplify default encoding logic

* run black on tests

* Don't force Accept header

* add test