Support datetimes in build queries
Previously, only `datetime.date` worked in all of the
`{created,finished}_{from,to}` arguments for build querying. If you
provided a `datetime.datetime` for any of these arguments, the `str`
function would be used for the `datetime.datetime`, and you'd end up
with a space separator: "YYYY-MM-DD HH:MM:SS".

It turns out that `datetime.date`'s str function returns a valid ISO
date string of "YYYY-MM-DD", but the `datetime.datetime` str function
uses a space separator instead of `T`.